requirements.txt
pyperclip==1.8.0
pycryptodome==3.10.1
